+++ src/BoolField/BoolFieldServiceProvider.php
@@ -216,30 +216,6 @@ final readonly class BoolFieldServiceProvider implements ServiceProviderInterfac
                 },
             )
             ->set(
-                FieldKeyNotRegisteredExceptionHandler::class,
-                static function (ContainerInterface $container): FieldKeyNotRegisteredExceptionHandler {
-                    $problemDetails = $container->get(ProblemDetailsResponseFactory::class);
-
-                    if (!$problemDetails instanceof ProblemDetailsResponseFactory) {
-                        throw new LogicException('Problem details response factory service is invalid.');
-                    }
-
-                    return new FieldKeyNotRegisteredExceptionHandler($problemDetails);
-                },
-            )
-            ->set(
-                FieldTypeMismatchExceptionHandler::class,
-                static function (ContainerInterface $container): FieldTypeMismatchExceptionHandler {
-                    $problemDetails = $container->get(ProblemDetailsResponseFactory::class);
-
-                    if (!$problemDetails instanceof ProblemDetailsResponseFactory) {
-                        throw new LogicException('Problem details response factory service is invalid.');
-                    }
-
-                    return new FieldTypeMismatchExceptionHandler($problemDetails);
-                },
-            )
-            ->set(
                 'nene-records.route_registrar.bool_field',
                 static function (ContainerInterface $container): BoolFieldRouteRegistrar {
                     $list = $container->get(ListBoolFieldsHandler::class);
