--- src/DateTimeField/FieldKeyNotRegisteredException.php
@@ -1,17 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace NeNeRecords\DateTimeField;
-
-use RuntimeException;
-
-final class FieldKeyNotRegisteredException extends RuntimeException
-{
-    public function __construct(
-        public int $entityTypeId,
-        public string $fieldKey,
-    ) {
-        parent::__construct("Field key {$fieldKey} is not registered for entity type {$entityTypeId}.");
-    }
-}
