--- src/DateTimeField/FieldTypeMismatchExceptionHandler.php
@@ -1,35 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace NeNeRecords\DateTimeField;
-
-use Nene2\Error\DomainExceptionHandlerInterface;
-use Nene2\Error\ProblemDetailsResponseFactory;
-use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\ServerRequestInterface;
-use Throwable;
-
-final readonly class FieldTypeMismatchExceptionHandler implements DomainExceptionHandlerInterface
-{
-    public function __construct(
-        private ProblemDetailsResponseFactory $problemDetails,
-    ) {
-    }
-
-    public function supports(Throwable $exception): bool
-    {
-        return $exception instanceof FieldTypeMismatchException;
-    }
-
-    public function handle(Throwable $exception, ServerRequestInterface $request): ResponseInterface
-    {
-        return $this->problemDetails->create(
-            $request,
-            'field-type-mismatch',
-            'Field Type Mismatch',
-            422,
-            'The field key is registered with a different data type.',
-        );
-    }
-}
