+++ src/EntityRelation/ListEntityRelationsUseCase.php
@@ -7,6 +7,8 @@ namespace NeNeRecords\EntityRelation;
 use NeNeRecords\Entity\EntityNotFoundException;
 use NeNeRecords\Entity\EntityRepositoryInterface;
 use NeNeRecords\FieldDef\FieldDefRepositoryInterface;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredException;
+use NeNeRecords\FieldDef\FieldTypeMismatchException;
 
 final readonly class ListEntityRelationsUseCase implements ListEntityRelationsUseCaseInterface
 {
