+++ src/TextField/CreateTextFieldUseCase.php
@@ -7,6 +7,8 @@ namespace NeNeRecords\TextField;
 use NeNeRecords\Entity\EntityNotFoundException;
 use NeNeRecords\Entity\EntityRepositoryInterface;
 use NeNeRecords\FieldDef\FieldDefRepositoryInterface;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredException;
+use NeNeRecords\FieldDef\FieldTypeMismatchException;
 
 final readonly class CreateTextFieldUseCase implements CreateTextFieldUseCaseInterface
 {
@@ -50,7 +52,7 @@ final readonly class CreateTextFieldUseCase implements CreateTextFieldUseCaseInt
         $fieldDef = $this->fieldDefs->findByEntityTypeIdAndFieldKey($entityTypeId, $fieldKey);
 
         if ($fieldDef === null) {
-            throw new FieldKeyNotRegisteredException($entityTypeId, $fieldKey);
+            throw new FieldKeyNotRegisteredException($fieldKey, $entityTypeId);
         }
 
         if ($fieldDef->dataType !== self::TEXT_DATA_TYPE) {
