+++ src/TextField/UpdateTextFieldUseCase.php
@@ -7,6 +7,8 @@ namespace NeNeRecords\TextField;
 use NeNeRecords\Entity\EntityNotFoundException;
 use NeNeRecords\Entity\EntityRepositoryInterface;
 use NeNeRecords\FieldDef\FieldDefRepositoryInterface;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredException;
+use NeNeRecords\FieldDef\FieldTypeMismatchException;
 
 final readonly class UpdateTextFieldUseCase implements UpdateTextFieldUseCaseInterface
 {
@@ -58,7 +60,7 @@ final readonly class UpdateTextFieldUseCase implements UpdateTextFieldUseCaseInt
         $fieldDef = $this->fieldDefs->findByEntityTypeIdAndFieldKey($entityTypeId, $fieldKey);
 
         if ($fieldDef === null) {
-            throw new FieldKeyNotRegisteredException($entityTypeId, $fieldKey);
+            throw new FieldKeyNotRegisteredException($fieldKey, $entityTypeId);
         }
 
         if ($fieldDef->dataType !== self::TEXT_DATA_TYPE) {
