+++ tests/BoolField/BoolFieldHttpTest.php
@@ -13,8 +13,6 @@ use NeNeRecords\BoolField\CreateBoolFieldHandler;
 use NeNeRecords\BoolField\CreateBoolFieldUseCase;
 use NeNeRecords\BoolField\DeleteBoolFieldHandler;
 use NeNeRecords\BoolField\DeleteBoolFieldUseCase;
-use NeNeRecords\BoolField\FieldKeyNotRegisteredExceptionHandler;
-use NeNeRecords\BoolField\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\BoolField\GetBoolFieldByIdHandler;
 use NeNeRecords\BoolField\GetBoolFieldByIdUseCase;
 use NeNeRecords\BoolField\ListBoolFieldsHandler;
@@ -45,6 +43,8 @@ use NeNeRecords\Entity\UpdateEntityUseCase;
 use NeNeRecords\EntityType\EntityType;
 use NeNeRecords\EntityType\EntityTypeNotFoundExceptionHandler;
 use NeNeRecords\FieldDef\FieldDef;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredExceptionHandler;
+use NeNeRecords\FieldDef\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\Tests\Entity\InMemoryEntityRepository;
 use NeNeRecords\Tests\EntityType\InMemoryEntityTypeRepository;
 use NeNeRecords\Tests\FieldDef\InMemoryFieldDefRepository;
