+++ tests/BoolField/CreateBoolFieldUseCaseTest.php
@@ -7,12 +7,12 @@ namespace NeNeRecords\Tests\BoolField;
 use NeNeRecords\BoolField\BoolField;
 use NeNeRecords\BoolField\CreateBoolFieldInput;
 use NeNeRecords\BoolField\CreateBoolFieldUseCase;
-use NeNeRecords\BoolField\FieldKeyNotRegisteredException;
-use NeNeRecords\BoolField\FieldTypeMismatchException;
 use NeNeRecords\BoolField\UpdateBoolFieldInput;
 use NeNeRecords\BoolField\UpdateBoolFieldUseCase;
 use NeNeRecords\Entity\Entity;
 use NeNeRecords\FieldDef\FieldDef;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredException;
+use NeNeRecords\FieldDef\FieldTypeMismatchException;
 use NeNeRecords\Tests\Entity\InMemoryEntityRepository;
 use NeNeRecords\Tests\FieldDef\InMemoryFieldDefRepository;
 use PHPUnit\Framework\TestCase;
