+++ tests/EntityRelation/EntityRelationHttpTest.php
@@ -14,14 +14,14 @@ use NeNeRecords\EntityRelation\AttachEntityRelationUseCase;
 use NeNeRecords\EntityRelation\DetachEntityRelationHandler;
 use NeNeRecords\EntityRelation\DetachEntityRelationUseCase;
 use NeNeRecords\EntityRelation\EntityRelationRouteRegistrar;
-use NeNeRecords\EntityRelation\FieldKeyNotRegisteredExceptionHandler;
-use NeNeRecords\EntityRelation\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\EntityRelation\ListEntityRelationsHandler;
 use NeNeRecords\EntityRelation\ListEntityRelationsUseCase;
 use NeNeRecords\EntityRelation\RelationAlreadyAttachedExceptionHandler;
 use NeNeRecords\EntityRelation\RelationNotAttachedExceptionHandler;
 use NeNeRecords\EntityRelation\RelationTargetTypeMismatchExceptionHandler;
 use NeNeRecords\FieldDef\FieldDef;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredExceptionHandler;
+use NeNeRecords\FieldDef\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\Tests\Entity\InMemoryEntityRepository;
 use NeNeRecords\Tests\FieldDef\InMemoryFieldDefRepository;
 use Nyholm\Psr7\Factory\Psr17Factory;
