+++ tests/EnumField/CreateEnumFieldUseCaseTest.php
@@ -8,11 +8,11 @@ use NeNeRecords\Entity\Entity;
 use NeNeRecords\EnumField\CreateEnumFieldInput;
 use NeNeRecords\EnumField\CreateEnumFieldUseCase;
 use NeNeRecords\EnumField\EnumField;
-use NeNeRecords\EnumField\FieldKeyNotRegisteredException;
-use NeNeRecords\EnumField\FieldTypeMismatchException;
 use NeNeRecords\EnumField\UpdateEnumFieldInput;
 use NeNeRecords\EnumField\UpdateEnumFieldUseCase;
 use NeNeRecords\FieldDef\FieldDef;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredException;
+use NeNeRecords\FieldDef\FieldTypeMismatchException;
 use NeNeRecords\Tests\Entity\InMemoryEntityRepository;
 use NeNeRecords\Tests\FieldDef\InMemoryFieldDefRepository;
 use PHPUnit\Framework\TestCase;
