+++ tests/EnumField/EnumFieldHttpTest.php
@@ -36,8 +36,6 @@ use NeNeRecords\EnumField\DeleteEnumFieldHandler;
 use NeNeRecords\EnumField\DeleteEnumFieldUseCase;
 use NeNeRecords\EnumField\EnumFieldNotFoundExceptionHandler;
 use NeNeRecords\EnumField\EnumFieldRouteRegistrar;
-use NeNeRecords\EnumField\FieldKeyNotRegisteredExceptionHandler;
-use NeNeRecords\EnumField\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\EnumField\GetEnumFieldByIdHandler;
 use NeNeRecords\EnumField\GetEnumFieldByIdUseCase;
 use NeNeRecords\EnumField\ListEnumFieldsHandler;
@@ -45,6 +43,8 @@ use NeNeRecords\EnumField\ListEnumFieldsUseCase;
 use NeNeRecords\EnumField\UpdateEnumFieldHandler;
 use NeNeRecords\EnumField\UpdateEnumFieldUseCase;
 use NeNeRecords\FieldDef\FieldDef;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredExceptionHandler;
+use NeNeRecords\FieldDef\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\Tests\Entity\InMemoryEntityRepository;
 use NeNeRecords\Tests\EntityType\InMemoryEntityTypeRepository;
 use NeNeRecords\Tests\FieldDef\InMemoryFieldDefRepository;
