+++ tests/IntField/IntFieldHttpTest.php
@@ -31,12 +31,12 @@ use NeNeRecords\Entity\UpdateEntityUseCase;
 use NeNeRecords\EntityType\EntityType;
 use NeNeRecords\EntityType\EntityTypeNotFoundExceptionHandler;
 use NeNeRecords\FieldDef\FieldDef;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredExceptionHandler;
+use NeNeRecords\FieldDef\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\IntField\CreateIntFieldHandler;
 use NeNeRecords\IntField\CreateIntFieldUseCase;
 use NeNeRecords\IntField\DeleteIntFieldHandler;
 use NeNeRecords\IntField\DeleteIntFieldUseCase;
-use NeNeRecords\IntField\FieldKeyNotRegisteredExceptionHandler;
-use NeNeRecords\IntField\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\IntField\GetIntFieldByIdHandler;
 use NeNeRecords\IntField\GetIntFieldByIdUseCase;
 use NeNeRecords\IntField\IntFieldNotFoundExceptionHandler;
