+++ tests/TextField/TextFieldHttpTest.php
@@ -31,6 +31,8 @@ use NeNeRecords\Entity\UpdateEntityUseCase;
 use NeNeRecords\EntityType\EntityType;
 use NeNeRecords\EntityType\EntityTypeNotFoundExceptionHandler;
 use NeNeRecords\FieldDef\FieldDef;
+use NeNeRecords\FieldDef\FieldKeyNotRegisteredExceptionHandler;
+use NeNeRecords\FieldDef\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\Tests\Entity\InMemoryEntityRepository;
 use NeNeRecords\Tests\EntityType\InMemoryEntityTypeRepository;
 use NeNeRecords\Tests\FieldDef\InMemoryFieldDefRepository;
@@ -38,8 +40,6 @@ use NeNeRecords\TextField\CreateTextFieldHandler;
 use NeNeRecords\TextField\CreateTextFieldUseCase;
 use NeNeRecords\TextField\DeleteTextFieldHandler;
 use NeNeRecords\TextField\DeleteTextFieldUseCase;
-use NeNeRecords\TextField\FieldKeyNotRegisteredExceptionHandler;
-use NeNeRecords\TextField\FieldTypeMismatchExceptionHandler;
 use NeNeRecords\TextField\GetTextFieldByIdHandler;
 use NeNeRecords\TextField\GetTextFieldByIdUseCase;
 use NeNeRecords\TextField\ListTextFieldsHandler;
